trip/requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
bunch==1.0.1
celery==4.4.7
Django==3.1.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-redis==4.12.1
kombu==4.6.11
mysqlclient==1.4.6
Pillow==7.2.0
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
vine==1.3.0
